changed dict indices to uint8
(cherry picked from commit c018336609ed92728f4bebf6d9e661f372ac9b0b)

+++ rust/lance-encoding/src/encoder.rs
@@ -247,7 +247,7 @@ impl CoreArrayEncodingStrategy {
             DataType::Utf8 | DataType::LargeUtf8 | DataType::Binary | DataType::LargeBinary => {
                 if use_dict_encoding {
                     let dict_indices_encoder =
-                        Self::array_encoder_from_type(&DataType::UInt64, false)?;
+                        Self::array_encoder_from_type(&DataType::UInt8, false)?;
                     let dict_items_encoder = Self::array_encoder_from_type(&DataType::Utf8, false)?;
 
                     Ok(Box::new(DictionaryEncoder::new(
